pulled cvodes_ode_data into distinct file and enabled all tests again
