+++ stan/math/rev/mat.hpp
@@ -53,6 +53,8 @@
 #include <stan/math/rev/mat/functor/gradient.hpp>
 #include <stan/math/rev/mat/functor/jacobian.hpp>
 #include <stan/math/rev/mat/functor/ode_system.hpp>
+#include <stan/math/rev/mat/functor/cvodes_utils.hpp>
+#include <stan/math/rev/mat/functor/cvodes_ode_data.hpp>
 #include <stan/math/rev/mat/functor/integrate_ode_bdf.hpp>
 
 #endif
