+++ test/unit/math/rev/mat/functor/cvodes_ode_data_prim_test.cpp
@@ -8,17 +8,15 @@
 struct StanMathOdeCVode : public ::testing::Test {
   void SetUp() {
     stan::math::recover_memory();
-    ts = std::vector<double>(1,10);
     t0 = 0;
   }
   std::stringstream msgs;
   std::vector<double> x;
   std::vector<int> x_int;
   double t0;
-  std::vector<double> ts;
 };
 TEST_F(StanMathOdeCVode, recover_exception) {
-  using stan::math::cvodes_integrator;
+  using stan::math::cvodes_ode_data;
   std::string message = "ode throws";
 
   const int N = 3;
@@ -29,8 +27,8 @@ TEST_F(StanMathOdeCVode, recover_exception) {
   std::vector<double> y0_d(N, 0.0);
   std::vector<double> theta_v(M, 0.0);
 
-  cvodes_integrator<mock_throwing_ode_functor<std::logic_error>, double, double>
-    integrator_dd(throwing_ode, y0_d, t0, theta_v, x, x_int, ts, 1e-8, 1e-10, 1e6, 1, &msgs);
+  cvodes_ode_data<mock_throwing_ode_functor<std::logic_error>, double, double>
+    ode_data_dd(throwing_ode, y0_d, theta_v, x, x_int, &msgs);
 
   std::vector<double> y(3,0);
   std::vector<double> dy_dt(3,0);
@@ -41,7 +39,7 @@ TEST_F(StanMathOdeCVode, recover_exception) {
 
   double t = 10;
 
-  EXPECT_THROW_MSG(integrator_dd.ode_rhs(t, nv_y, nv_dy_dt, &integrator_dd),
+  EXPECT_THROW_MSG(ode_data_dd.ode_rhs(t, nv_y, nv_dy_dt, &ode_data_dd),
                    std::logic_error,
                    message);
 }
